requirements.txt
absl-py==0.10.0
aiocontextvars==0.2.2
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextvars==2.4
dataclasses==0.7
fastapi==0.58.0
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==2.0.0
joblib==0.15.1
Keras-Preprocessing==1.1.2
loguru==0.5.1
Markdown==3.2.2
multidict==4.7.6
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.5.1
python-multipart==0.0.5
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
starlette==0.13.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.7
uvloop==0.14.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
